fix crud in correct typescript and an integrity control

+++ services/product/src/handler/Allergen/delete.ts
@@ -1,22 +1,25 @@
-import { AllergenId } from "../../types/Allergen";
-import { Data } from "../../types";
-import { prisma } from "../../lib/prisma";
-import { log } from "../../lib/log";
+import { Allergen, AllergenId } from "@product/types/Allergen";
+import { Data } from "@product/types";
+import { log } from "@product/lib/log";
+import { ServerErrorResponse } from "@grpc/grpc-js";
+import prisma from "@product/lib/prisma";
 
 export const DeleteAllergen = async (
 	data: Data<AllergenId>,
-	callback: (err: any, response: any) => void
+	callback: (err: ServerErrorResponse | null) => void
 ) => {
 	log.debug("Request received at DeleteAllergen handler\n", data.request);
 	try {
-		const { request } = data;
-		const { id } = request;
+		const { id } = data.request;
+
+		if((await prisma.allergen.findFirst({where: {id: id}}) as Allergen | null) == null)
+			throw(Error("L'allergen n'existe  pas") as ServerErrorResponse)
 
 		await prisma.allergen.delete({ where : {id} });
 
-		callback(null, null);
-	} catch (error) {
+		callback(null);
+	} catch (error: ServerErrorResponse | any) {
 		log.error(error);
-		callback(error, null);
+		callback(error);
 	}
 };
